Expose `evcxr_jupyter` as a library
